requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==1.2.3
    # via jinja2-time
binaryornot==0.4.4
    # via cookiecutter
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
chardet==5.0.0
    # via binaryornot
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   cookiecutter
    #   dask
    #   distributed
    #   flytekit
cloudpickle==2.2.0
    # via
    #   dask
    #   distributed
    #   flytekit
cookiecutter==2.1.1
    # via flytekit
croniter==1.3.7
    # via flytekit
cryptography==38.0.3
    # via pyopenssl
dask[distributed]==2022.10.2
    # via
    #   distributed
    #   flytekitplugins-dask
dataclasses-json==0.5.7
    # via flytekit
decorator==5.1.1
    # via retry
deprecated==1.2.13
    # via flytekit
diskcache==5.4.0
    # via flytekit
distributed==2022.10.2
    # via dask
docker==6.0.1
    # via flytekit
docker-image-py==0.1.12
    # via flytekit
docstring-parser==0.15
    # via flytekit
flyteidl==1.3.2
    # via
    #   flytekit
    #   flytekitplugins-dask
flytekit==1.3.1
    # via flytekitplugins-dask
flytekitplugins-dask==1.3.1
    # via -r requirements.in
fsspec==2022.10.0
    # via dask
gitdb==4.0.10
    # via gitpython
gitpython==3.1.30
    # via flytekit
googleapis-common-protos==1.58.0
    # via
    #   flyteidl
    #   flytekit
    #   grpcio-status
grpcio==1.51.1
    # via
    #   flytekit
    #   grpcio-status
grpcio-status==1.51.1
    # via flytekit
heapdict==1.0.1
    # via zict
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via
    #   flytekit
    #   keyring
jaraco-classes==3.2.3
    # via keyring
jinja2==3.1.2
    # via
    #   cookiecutter
    #   distributed
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
joblib==1.2.0
    # via flytekit
keyring==23.11.0
    # via flytekit
locket==1.0.0
    # via
    #   distributed
    #   partd
markupsafe==2.1.1
    # via jinja2
marshmallow==3.18.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via flytekit
more-itertools==9.0.0
    # via jaraco-classes
msgpack==1.0.4
    # via distributed
mypy-extensions==0.4.3
    # via typing-inspect
natsort==8.2.0
    # via flytekit
numpy==1.23.4
    # via
    #   flytekit
    #   pandas
    #   pyarrow
    #   scipy
packaging==21.3
    # via
    #   dask
    #   distributed
    #   docker
    #   marshmallow
pandas==1.5.1
    # via flytekit
partd==1.3.0
    # via dask
protobuf==4.21.11
    # via
    #   flyteidl
    #   googleapis-common-protos
    #   grpcio-status
    #   protoc-gen-swagger
protoc-gen-swagger==0.1.0
    # via flyteidl
psutil==5.9.3
    # via distributed
py==1.11.0
    # via retry
pyarrow==6.0.1
    # via flytekit
pycparser==2.21
    # via cffi
pyopenssl==22.1.0
    # via flytekit
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   flytekit
    #   pandas
python-json-logger==2.0.4
    # via flytekit
python-slugify==6.1.2
    # via cookiecutter
pytimeparse==1.1.8
    # via flytekit
pytz==2022.6
    # via
    #   flytekit
    #   pandas
pyyaml==6.0
    # via
    #   cookiecutter
    #   dask
    #   distributed
    #   flytekit
regex==2022.10.31
    # via docker-image-py
requests==2.28.1
    # via
    #   cookiecutter
    #   docker
    #   flytekit
    #   responses
responses==0.22.0
    # via flytekit
retry==0.9.2
    # via flytekit
scipy==1.10.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
sortedcontainers==2.4.0
    # via
    #   distributed
    #   flytekit
statsd==3.3.0
    # via flytekit
tblib==1.7.0
    # via distributed
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via responses
toolz==0.12.0
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via distributed
types-toml==0.10.8
    # via responses
typing-extensions==4.4.0
    # via
    #   flytekit
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
urllib3==1.26.12
    # via
    #   distributed
    #   docker
    #   flytekit
    #   requests
    #   responses
websocket-client==1.4.2
    # via docker
wheel==0.38.2
    # via flytekit
wrapt==1.14.1
    # via
    #   deprecated
    #   flytekit
zict==2.2.0
    # via distributed
zipp==3.10.0
    # via importlib-metadata
